I changed something, i think

+++ ft_bzero.c
@@ -6,13 +6,15 @@
 /*   By: hbreeze <hbreeze@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/27 18:09:58 by hbreeze           #+#    #+#             */
-/*   Updated: 2024/08/28 15:05:50 by hbreeze          ###   ########.fr       */
+/*   Updated: 2024/08/30 13:05:42 by hbreeze          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*t;
 
 	t = s;
