+++ ft_memcmp.c
@@ -6,15 +6,17 @@
 /*   By: hbreeze <hbreeze@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/27 20:03:19 by hbreeze           #+#    #+#             */
-/*   Updated: 2024/08/28 15:06:13 by hbreeze          ###   ########.fr       */
+/*   Updated: 2024/08/30 13:06:45 by hbreeze          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	const unsigned char	*s1_i;
 	const unsigned char	*s2_i;
-	unsigned int		index;
+	size_t				index;
 
 	s1_i = s1;
 	s2_i = s2;
