+++ ft_striteri.c
@@ -6,13 +6,15 @@
 /*   By: hbreeze <hbreeze@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/28 13:05:57 by hbreeze           #+#    #+#             */
-/*   Updated: 2024/08/28 13:13:41 by hbreeze          ###   ########.fr       */
+/*   Updated: 2024/08/30 13:08:20 by hbreeze          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	unsigned int	index;
+	size_t	index;
 
 	index = 0;
 	if (!s || !f)
