+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: hbreeze <hbreeze@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/28 10:50:51 by hbreeze           #+#    #+#             */
-/*   Updated: 2024/08/28 20:18:17 by hbreeze          ###   ########.fr       */
+/*   Updated: 2024/08/30 13:04:31 by hbreeze          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 char	*ft_strjoin(const char *s1, const char *s2)
 {
 	char			*output;
-	unsigned int	size;
-	unsigned int	index;
+	size_t			size;
+	size_t			index;
 
 	if (s1 == 0 || s2 == 0)
 		return (0);
