+++ ft_strlcat.c
@@ -6,17 +6,17 @@
 /*   By: hbreeze <hbreeze@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/27 18:33:35 by hbreeze           #+#    #+#             */
-/*   Updated: 2024/08/29 13:21:05 by hbreeze          ###   ########.fr       */
+/*   Updated: 2024/08/30 12:59:08 by hbreeze          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
 	char			*dest_i;
-	unsigned int	dest_len;
-	unsigned int	src_len;
+	size_t			dest_len;
+	size_t			src_len;
 
 	if (!dest || !src)
 		return (0);
